requirements.txt
accelerate>=0.25.0
librosa>=0.10.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
safetensors>=0.4.0
soundfile>=0.12.1
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.38.0
yt-dlp>=2024.3.10
google-genai